+++ src/vuetifyConfig.ts
@@ -1,0 +1,54 @@
+import { createVuetify } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+import { aliases, mdi } from 'vuetify/iconsets/mdi-svg'
+import {
+	cnamColorsTokens,
+	cnamContextualTokens,
+	cnamLightTheme,
+	cnamDarkTheme,
+	paColorsTokens,
+	paContextualTokens,
+	paLightTheme,
+	paDarkTheme,
+} from './designTokens'
+import { createFlattenTheme } from './designTokens/utils'
+
+export const createVuetifyInstance = () => createVuetify({
+	components,
+	directives,
+	theme: {
+		defaultTheme: 'cnam',
+		themes: {
+			cnam: {
+				dark: false,
+				colors: {
+					...cnamLightTheme,
+					...cnamDarkTheme,
+				},
+				variables: {
+					'border-color': cnamColorsTokens.grey.base,
+					...createFlattenTheme(cnamContextualTokens),
+				},
+			},
+			pa: {
+				dark: false,
+				colors: {
+					...paLightTheme,
+					...paDarkTheme,
+				},
+				variables: {
+					'border-color': paColorsTokens.grey.base,
+					...createFlattenTheme(paContextualTokens),
+				},
+			},
+		},
+	},
+	icons: {
+		defaultSet: 'mdi',
+		aliases,
+		sets: {
+			mdi,
+		},
+	},
+})
